API

DataVisualization.UI — Type

UI(table, tabs=(:Load, :Filter, :Predict, :Cluster, :Project, :Visualize))

Generate a UI with a given table as starting value. tabs denote the list of tabs to include in the user interface. Each tab can take one of the following types: :Load, :Filter, :Predict, :Cluster, :Project, :Visualize. Repetitions are allowed, for example setting tabs=(:Load, :Filter, :Predict, :Filter, :Visualize) would generate a UI that allowes filtering both before and after fitting a linear model to the data.

DataVisualization.app — Function

app(table, tabs=(:Load, :Filter, :Predict, :Cluster, :Project, :Visualize))

Generate a UI with a given table and list of tabs. Launch the output as a local app.

DataVisualization.serve — Function

serve(table, tabs=(:Load, :Filter, :Predict, :Cluster, :Project, :Visualize);
      url=Sockets.localhost, port=8081)

Generate a UI with a given table and list of tabs. Serve the output at the given url and port. Return a JSServe.Server object.